- Fix compilation with MinGW + w32api and clean up warnings.
SVN r2505 (trunk)

+++ src/sound/music_mus_midiout.cpp
@@ -125,7 +125,7 @@ MUSSong2::MUSSong2 (FILE *file, BYTE *musiccache, int len, EMIDIDevice type)
 
 	// Read the remainder of the song.
 	len = int(len - start);
-	if (len < sizeof(MusHeader))
+	if (len < (int)sizeof(MusHeader))
 	{ // It's too short.
 		return;
 	}
